+++ packages/scripts/CHANGELOG.md
@@ -1,3 +1,10 @@
+# newspack-scripts [5.11.0](https://github.com/Automattic/newspack-workspace/compare/newspack-scripts@5.10.0...newspack-scripts@5.11.0) (2026-08-17)
+
+
+### Features
+
+* **content-gate:** grant paid access from one-time purchases with an access duration (NPPD-2053, [#693](https://github.com/Automattic/newspack-workspace/issues/693)) ([ae041fe](https://github.com/Automattic/newspack-workspace/commit/ae041fe123bce498dc6c98ab1247d2fcbcb74ba3))
+
 # newspack-scripts [5.10.0](https://github.com/Automattic/newspack-workspace/compare/newspack-scripts@5.9.8...newspack-scripts@5.10.0) (2026-08-03)